requirements.txt
# Core
streamlit>=1.40.0
python-dotenv>=1.0.0

# PyTorch (required for sentence-transformers)
torch>=2.0.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Embeddings & Vector Store
sentence-transformers>=3.0.0
chromadb>=1.0.0

# Document processing
pypdf>=4.0.0
pdfplumber>=0.11.0
pandas>=2.0.0

# HuggingFace
huggingface-hub>=0.24.0
transformers>=4.46.0
tokenizers>=0.20.0

# Other
tiktoken>=0.7.0
openai>=1.0.0

# Pydantic (required by chromadb)
pydantic>=2.0.0
pydantic-settings>=2.0.0
